Separate data options classes into three files
