+++ src/mlpack/core/data/data_options.hpp
@@ -1,5 +1,5 @@
 /**
- * @file core/data/load_impl.hpp
+ * @file core/data/data_options.hpp
  * @author Ryan Curtin
  * @author Omar Shrit
  *
@@ -213,367 +213,6 @@ class EmptyOptions : public DataOptionsBase<EmptyOptions>
 
 using DataOptions = DataOptionsBase<EmptyOptions>;
 
-template<typename Derived>
-class MatrixOptionsBase : public DataOptionsBase<MatrixOptionsBase<Derived>>
-{
- public:
-  MatrixOptionsBase(bool noTranspose = defaultNoTranspose) :
-      DataOptionsBase<MatrixOptionsBase<Derived>>(),
-      noTranspose(noTranspose)
-  {
-    // Do Nothing.
-  }
-
-  template<typename Derived2>
-  explicit MatrixOptionsBase(const MatrixOptionsBase<Derived2>& opts) :
-      DataOptionsBase<MatrixOptionsBase<Derived>>()
-  {
-    // Delegate to copy operator.
-    *this = opts;
-  }
-
-  template<typename Derived2>
-  explicit MatrixOptionsBase(MatrixOptionsBase<Derived2>&& opts) :
-      DataOptionsBase<MatrixOptionsBase<Derived>>()
-  {
-    // Delegate to move operator.
-    *this = std::move(opts);
-  }
-
-  // Inherit base class constructors.
-  using DataOptionsBase<MatrixOptionsBase<Derived>>::DataOptionsBase;
-
-  MatrixOptionsBase& operator=(const MatrixOptionsBase& other)
-  {
-    if (&other == this)
-      return *this;
-
-    if (other.noTranspose.has_value())
-      noTranspose = other.noTranspose;
-    DataOptionsBase<MatrixOptionsBase<Derived>>::CopyOptions(other);
-
-    return *this;
-  }
-
-  MatrixOptionsBase& operator=(MatrixOptionsBase&& other)
-  {
-    if (&other == this)
-      return *this;
-
-    noTranspose = std::move(other.noTranspose);
-    DataOptionsBase<MatrixOptionsBase<Derived>>::MoveOptions(std::move(other));
-
-    return *this;
-  }
-
-  template<typename Derived2>
-  MatrixOptionsBase& operator=(const MatrixOptionsBase<Derived2>& other)
-  {
-    if ((void*) &other == (void*) this)
-      return *this;
-
-    // Print warnings for any members that cannot be converted.
-    const char* dataDesc = static_cast<const Derived&>(*this).DataDescription();
-    static_cast<const Derived2&>(other).WarnBaseConversion(dataDesc);
-
-    // Only copy options that have been set in the other object.
-    if (other.noTranspose.has_value())
-      noTranspose = other.NoTranspose();
-
-    // Copy base members.
-    DataOptionsBase<MatrixOptionsBase<Derived>>::CopyOptions(other);
-
-    return *this;
-  }
-
-  template<typename Derived2>
-  MatrixOptionsBase& operator=(MatrixOptionsBase<Derived2>&& other)
-  {
-    if ((void*) this == (void*) &other)
-      return *this;
-
-    // Print warnings for any members that cannot be converted.
-    const char* dataDesc = static_cast<const Derived&>(*this).DataDescription();
-    static_cast<const Derived2&>(other).WarnBaseConversion(dataDesc);
-
-    noTranspose = std::move(other.noTranspose);
-
-    // Move base members.
-    DataOptionsBase<MatrixOptionsBase<Derived>>::MoveOptions(std::move(other));
-
-    return *this;
-  }
-
-  void WarnBaseConversion(const char* dataDescription) const
-  {
-    if (noTranspose.has_value() && noTranspose != defaultNoTranspose)
-      this->WarnOptionConversion("noTranspose", dataDescription);
-
-    // We may potentially need to print warnings for any other converted members
-    // of the derived type.
-    static_cast<const Derived&>(*this).WarnBaseConversion(dataDescription);
-  }
-
-  static const char* DataDescription() { return "matrix data"; }
-
-  void Reset()
-  {
-    noTranspose.reset();
-
-    // Reset any child members.
-    static_cast<Derived&>(*this).Reset();
-  }
-
-  // Get whether or not we will avoid transposing the matrix during load.
-  bool NoTranspose() const
-  {
-    return this->AccessMember(noTranspose, defaultNoTranspose);
-  }
-  // Modify whether or not we will avoid transposing the matrix during load.
-  bool& NoTranspose()
-  {
-    return this->ModifyMember(noTranspose, defaultNoTranspose);
-  }
-
- private:
-  std::optional<bool> noTranspose;
-
-  constexpr static const bool defaultNoTranspose = false;
-
-  // For access to internal optional members.
-  template<typename Derived2>
-  friend class MatrixOptionsBase;
-};
-
-// This utility class is meant to be used as the Derived parameter for a matrix
-// option that is not actually a derived type.  It provides the
-// WarnBaseConversion() member, which does nothing.
-class EmptyMatrixOptions : public MatrixOptionsBase<EmptyMatrixOptions>
-{
- public:
-  void WarnBaseConversion(const char* /* dataDescription */) const { }
-  static const char* DataDescription() { return "general data"; }
-  void Reset() { }
-};
-
-using MatrixOptions = MatrixOptionsBase<EmptyMatrixOptions>;
-
-class TextOptions : public MatrixOptionsBase<TextOptions>
-{
- public:
-  // TODO: pass through noTranspose option?
-  TextOptions(bool hasHeaders = defaultHasHeaders,
-              bool semicolon = defaultSemicolon,
-              bool missingToNan = defaultMissingToNan,
-              bool categorical = defaultCategorical,
-              bool missingPolicy = defaultMissingPolicy) :
-      MatrixOptionsBase<TextOptions>(),
-      hasHeaders(hasHeaders),
-      semicolon(semicolon),
-      missingToNan(missingToNan),
-      categorical(categorical),
-      missingPolicy(missingPolicy)
-  {
-    // Do Nothing.
-  }
-
-  explicit TextOptions(const TextOptions& opts) :
-      MatrixOptionsBase<TextOptions>()
-  {
-    // Delegate to copy operator.
-    *this = opts;
-  }
-
-  explicit TextOptions(TextOptions&& opts) :
-      MatrixOptionsBase<TextOptions>()
-  {
-    // Delegate to move operator.
-    *this = std::move(opts);
-  }
-
-  // Inherit base class constructors.
-  using MatrixOptionsBase<TextOptions>::MatrixOptionsBase;
-
-  TextOptions& operator=(const TextOptions& other)
-  {
-    if (&other == this)
-      return *this;
-
-    if (other.hasHeaders.has_value())
-      hasHeaders = *other.hasHeaders;
-    if (other.semicolon.has_value())
-      semicolon = *other.semicolon;
-    if (other.missingToNan.has_value())
-      missingToNan = *other.missingToNan;
-    if (other.categorical.has_value())
-      categorical = *other.categorical;
-    if (other.missingPolicy.has_value())
-      missingPolicy = *other.missingPolicy;
-
-    headers = other.headers;
-    datasetInfo = other.datasetInfo;
-    datasetMissingPolicy = other.datasetMissingPolicy;
-
-    // Copy base members.
-    MatrixOptionsBase<TextOptions>::operator=(other);
-
-    return *this;
-  }
-
-  TextOptions& operator=(TextOptions&& other)
-  {
-    if (&other == this)
-      return *this;
-
-    hasHeaders = std::move(other.hasHeaders);
-    semicolon = std::move(other.semicolon);
-    missingToNan = std::move(other.missingToNan);
-    categorical = std::move(other.categorical);
-    missingPolicy = std::move(other.missingPolicy);
-
-    headers = std::move(other.headers);
-    datasetInfo = std::move(other.datasetInfo);
-    datasetMissingPolicy = std::move(other.datasetMissingPolicy);
-
-    // Move base members.
-    MatrixOptionsBase<TextOptions>::operator=(std::move(other));
-
-    return *this;
-  }
-
-  // Print warnings for any members that cannot be represented by a
-  // DataOptionsBase<void>.
-  void WarnBaseConversion(const char* dataDescription) const
-  {
-    if (missingToNan.has_value() && missingToNan != defaultMissingToNan)
-      this->WarnOptionConversion("missingToNan", dataDescription);
-    if (semicolon.has_value() && semicolon != defaultSemicolon)
-      this->WarnOptionConversion("semicolon", dataDescription);
-    if (categorical.has_value() && categorical != defaultCategorical)
-      this->WarnOptionConversion("categorical", dataDescription);
-    if (hasHeaders.has_value() && hasHeaders != defaultHasHeaders)
-      this->WarnOptionConversion("hasHeaders", dataDescription);
-
-    // If either headers or datasetInfo are non-empty, then we take it that the
-    // user has manually modified them.
-    if (!headers.is_empty())
-      this->WarnOptionConversion("headers", dataDescription);
-    if (datasetInfo.Dimensionality() > 0)
-      this->WarnOptionConversion("datasetInfo", dataDescription);
-  }
-
-  static const char* DataDescription() { return "text-file matrix data"; }
-
-  void Reset()
-  {
-    hasHeaders.reset();
-    semicolon.reset();
-    missingToNan.reset();
-    categorical.reset();
-    missingPolicy.reset();
-  }
-
-  // Get if the dataset has headers or not.
-  bool HasHeaders() const
-  {
-    return this->AccessMember(hasHeaders, defaultHasHeaders);
-  }
-  // Modify if the dataset has headers.
-  bool& HasHeaders()
-  {
-    return this->ModifyMember(hasHeaders, defaultHasHeaders);
-  }
-
-  // Get if the separator is a semicolon in the data file.
-  bool Semicolon() const
-  {
-    return this->AccessMember(semicolon, defaultSemicolon);
-  }
-  // Modify the separator type in the matrix.
-  bool& SemiColon()
-  {
-    return this->ModifyMember(semicolon, defaultSemicolon);
-  }
-
-  // Get whether missing values are converted to NaN values.
-  bool MissingToNan() const
-  {
-    return this->AccessMember(missingToNan, defaultMissingToNan);
-  }
-  // Modify whether missing values are converted to NaN values.
-  bool& MissingToNan()
-  {
-    return this->ModifyMember(missingToNan, defaultMissingToNan);
-  }
-
-  // Get whether the data should be interpreted as categorical when columns are
-  // not numeric.
-  bool Categorical() const
-  {
-    return this->AccessMember(categorical, defaultCategorical);
-  }
-  // Modify whether the data should be interpreted as categorical when columns
-  // are not numeric.
-  bool& Categorical()
-  {
-    return this->ModifyMember(categorical, defaultCategorical);
-  }
-
-  // Get whether the data should be interpreted as categorical when columns are
-  // not numeric.
-  bool MissingPolicy() const
-  {
-    return this->AccessMember(missingPolicy, defaultMissingPolicy);
-  }
-  // Modify whether the data should be interpreted as missingPolicy when columns
-  // are not numeric.
-  bool& MissingPolicy()
-  {
-    return this->ModifyMember(missingPolicy, defaultMissingPolicy);
-  }
-
-  // Get the headers.
-  const arma::field<std::string>& Headers() const { return headers; }
-  // Modify the headers.
-  arma::field<std::string>& Headers() { return headers; }
-
-  // Get the DatasetInfo for categorical data.
-  const data::DatasetInfo& DatasetInfo() const { return datasetInfo; }
-  // Modify the DatasetInfo.
-  data::DatasetInfo& DatasetInfo() { return datasetInfo; }
-
-  // Get the DatasetInfo for categorical data.
-  const data::DatasetMapper<data::MissingPolicy>& DatasetMissingPolicy() const
-  { 
-    return datasetMissingPolicy;
-  }
-
-  // Modify the DatasetMissingPolicy.
-  data::DatasetMapper<data::MissingPolicy>& DatasetMissingPolicy()
-  {
-    return datasetMissingPolicy;
-  }
-
- private:
-  std::optional<bool> hasHeaders;
-  std::optional<bool> semicolon;
-  std::optional<bool> missingToNan;
-  std::optional<bool> categorical;
-  std::optional<bool> missingPolicy;
-
-  // These are not optional, but if either is specified, then it should be taken
-  // to mean that `hasHeaders` or `categorical` has been specified as true.
-  arma::field<std::string> headers;
-  data::DatasetInfo datasetInfo;
-  // Temporary internal member until MissingPolicy is refactored out.
-  data::DatasetMapper<data::MissingPolicy> datasetMissingPolicy;
-
-  constexpr static const bool defaultHasHeaders = false;
-  constexpr static const bool defaultSemicolon = false;
-  constexpr static const bool defaultMissingToNan = false;
-  constexpr static const bool defaultCategorical = false;
-  constexpr static const bool defaultMissingPolicy = false;
-};
 
 } // namespace data
 } // namespace mlpack
